+++ docs/Changelog.md
@@ -2,10 +2,11 @@
 # 6.0.0.1 Beta (not yet released)
 
 - New tutorial: [Extending mpv and mpv.net via Lua scripting](https://github.com/stax76/mpv.net/wiki/Extending-mpv-and-mpv.net-via-Lua-scripting)
-- New options `autofit-image` and `autofit-audio`, like autofit but used for image and audio files. Default 80.
+- New options `autofit-image` and `autofit-audio`, like autofit but used for image and audio files.
 - New [auto-mode](https://github.com/stax76/mpv-scripts) script to use mpv and mpv.net as image viewer and audio player.
 - New support of the mpv option `snap-window`.
 - New feature to show chapters in the command palette, see binding and menu definition below.
+- New option minimum-aspect-ratio-audio, same as minimum-aspect-ratio but used for audio files.
 - Fix long commands causing key bindings not visible in the command palette.
 - Fix script compatibility with mordenx and mpv-osc-tethys.
 - Fix borderless window not resizable with mouse.